move getItemStackFromValue to ValueConversions
